+++ permissions/src/epilot_permissions/models/getroleop.py
@@ -4,14 +4,16 @@ from __future__ import annotations
 from epilot_permissions.types import BaseModel
 from epilot_permissions.utils import FieldMetadata, PathParamMetadata
 import pydantic
-from typing import TypedDict
-from typing_extensions import Annotated
+from typing_extensions import Annotated, TypedDict
 
 
 class GetRoleRequestTypedDict(TypedDict):
     role_id: str
-    
+
 
 class GetRoleRequest(BaseModel):
-    role_id: Annotated[str, pydantic.Field(alias="roleId"), FieldMetadata(path=PathParamMetadata(style="simple", explode=False))]
-    
+    role_id: Annotated[
+        str,
+        pydantic.Field(alias="roleId"),
+        FieldMetadata(path=PathParamMetadata(style="simple", explode=False)),
+    ]